finished TodoApp

+++ server/server.js
@@ -2,18 +2,22 @@ const express = require('express')
 const bodyParser = require('body-parser')
 const _= require('lodash')
 const {ObjectID} = require('mongodb')
+const bcrypt = require('bcryptjs')
 
 const {mongoose} = require('./db/mongoose')
 const {Todo} = require('./models/Todo');
 const {User} = require('./models/User');
+const {authenticate} = require('./middleware/authenticate')
+const {config} = require('../server/config/config')
 
 const app = express();
 const port = process.env.PORT || 3000;
 app.use(bodyParser.json());
 
-app.post('/todos', (req,res)=>{
+app.post('/todos',authenticate, (req,res)=>{
    var todo = new Todo({
-       text: req.body.text
+       text: req.body.text,
+       _creator: req.user._id
    })
    todo.save().then((doc)=>{
     res.send(doc);
@@ -22,22 +26,24 @@ app.post('/todos', (req,res)=>{
    })
 })
 
-app.get('/todos',(req,res)=>{
-    Todo.find().then((todos)=>{
+app.get('/todos',authenticate, (req,res)=>{
+    Todo.find({
+        _creator: req.user._id
+    }).then((todos)=>{
         res.send({todos})
     },(err)=>{
         res.status(400).send(err);
     })
 })
 
-app.get('/todos/:id',(req,res)=>{
+app.get('/todos/:id',authenticate,(req,res)=>{
    var id = req.params.id
    if(!ObjectID.isValid(id))
    {
     return res.status(404).send()
    }
 
-    Todo.findById({_id: id}).then((todo)=>{
+    Todo.findById({_id: id, _creator: req.user._id}).then((todo)=>{
         if(!todo)
         {
             return res.status(404).send()
@@ -48,14 +54,14 @@ app.get('/todos/:id',(req,res)=>{
     })
 })
 
-app.delete('/todos/:id',(req,res)=>{
+app.delete('/todos/:id',authenticate,(req,res)=>{
     var id = req.params.id
     if(!ObjectID.isValid(id))
     {
      return res.status(404).send()
     }
  
-     Todo.findByIdAndRemove({_id: id}).then((todo)=>{
+     Todo.findByIdAndRemove({_id: id, _creator: req.user._id}).then((todo)=>{
          if(!todo)
          {
              return res.status(404).send()
@@ -66,7 +72,7 @@ app.delete('/todos/:id',(req,res)=>{
      })
  })
 
- app.patch('/todos/:id',(req,res)=>{
+ app.patch('/todos/:id',authenticate,(req,res)=>{
      var id = req.params.id;
      var body = _.pick(req.body, ['text', 'completed']);
      if(!ObjectID.isValid(id))
@@ -83,7 +89,7 @@ app.delete('/todos/:id',(req,res)=>{
         body.completedAt = null;
      }
 
-     Todo.findByIdAndUpdate(id,{$set:body}, {new: true}).then((todo)=>{
+     Todo.findByIdAndUpdate({_id:id,  _creator: req.user._id},{$set:body}, {new: true}).then((todo)=>{
         if(!todo)
         {
             return res.status(404).send()
@@ -96,7 +102,7 @@ app.delete('/todos/:id',(req,res)=>{
 
  //USERS
 
- app.post('/users',(req,res)=>{
+ app.post('/users/signup',(req,res)=>{
      var body = _.pick(req.body,['email','password']);
      var user = new User(body);
 
@@ -109,6 +115,30 @@ app.delete('/todos/:id',(req,res)=>{
          res.status(400).send(err)
      })
  })
+
+
+ app.get('/users/me',authenticate, (req,res)=>{
+    res.send(req.user);
+})
+
+app.post('/users/login',(req,res)=>{
+   var body = _.pick(req.body, ['email','password'])
+   console.log(body);
+    User.findByCred(body.email,body.password).then((user)=>{
+        return user.generateAuthToken().then((token)=>{
+            res.header('x-auth', token).send(user)
+        })
+    }).catch((err)=>{
+        res.status(400).send();
+    })
+})
+app.delete('/users/me/token',authenticate,(req,res)=>{
+    req.user.removeToken(req.token).then(()=>{
+        res.status(200).send()
+    },()=>{
+        res.status(400).send
+    })
+})
 app.listen(port, ()=>{
     console.log(`started on port ${port}`);
 })
